Allow configuring list of strict pod labels(#63)
This feature morphs the `WATCHER_STRICT_LABELS` configuration parameter from a boolean that enables restricting following pods labeled with `application` and `version` to a comma separated list of label names. This allows users to specify their own list of labels that need to be present to enable following of pods.