Add redaction rule for JWT tokens to all containers
JWT tokens must not be logged to Scalyr, so a redaction rule is added
to the Scalyr agent configuration section for each container. If no
custom redaction rules are set in the pod annotation, the JWT
redaction rule is the only rule. Otherwise the JWT redaction rule is
added to the end of the custom rules.

Signed-off-by: Christian Berg <berg.christian@gmail.com>